requirements.txt
argcomplete
python-tlsh
python-magic
